Show the channel tag and the number of channels in the program guide title.
This is the same behaviour as in the channel list activity. Also fixed
the amount of shown icons in the left side of the program guide. Some
optimizations were needed to get the desired result. The current tag is
now saved in the utils class and not separately in each class.