Remove right fragment in dual pane mode to avoid showing invalid data.
When the main fragment has added its list contents the onListPopulated
is called. Remove the right fragment in case the list in the main
fragment is empty to avoid showing previous data. If the main fragment
list contains data then the call to onListItemSelected would be made
witch would show the right fragment again.