Added variable that defines if all recordings should be reloaded after removal. Also access the app object only once.
The variable is required for the timer recording. When the user has
defined a recording without a channel (which should not be allowed by
the webinterface) and this recording shall be deleted, the server will
not confirm the deletion. In this case a manual reload is required to
fetch all updated data from the server again.